requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
requests==2.25.1
soupsieve==2.2.1
SQLAlchemy==1.4.3
urllib3==1.26.4
Werkzeug==1.0.1
python-dotenv==0.17.0
uwsgi==2.0.19.1
